Open notes in new window, instead of printing to console
It can be tedious to load up the speaker notes, with the whole having to
open the javascript console, and clicking on the link.

This simplifies it by automatically opening the notes in a new window.
Most browsers will warn that it's trying to open a pop-up, but it's easy
enough to allow it.